limit global variables to a minimun range
