+++ component/topsql/topsql.go
@@ -1,25 +1,49 @@
 package topsql
 
 import (
+	"github.com/gin-gonic/gin"
+	"github.com/pingcap/ng-monitoring/component/topsql/service"
 	"net/http"
 
+	"github.com/genjidb/genji"
 	"github.com/pingcap/ng-monitoring/component/topology"
 	"github.com/pingcap/ng-monitoring/component/topsql/query"
 	"github.com/pingcap/ng-monitoring/component/topsql/store"
 	"github.com/pingcap/ng-monitoring/component/topsql/subscriber"
 	"github.com/pingcap/ng-monitoring/config/pdvariable"
-
-	"github.com/genjidb/genji"
 )
 
-func Init(gj *genji.DB, insertHdr, selectHdr http.HandlerFunc, topSub topology.Subscriber, varSub pdvariable.Subscriber) {
-	store.Init(insertHdr, gj)
-	query.Init(selectHdr, gj)
-	subscriber.Init(topSub, varSub)
+var (
+	defStore      store.Store
+	defQuery      query.Query
+	defSubscriber subscriber.Subscriber
+	defService    *service.Service
+)
+
+func Init(
+	gj *genji.DB,
+	insertHdr, selectHdr http.HandlerFunc,
+	topSub topology.Subscriber,
+	varSub pdvariable.Subscriber,
+) (err error) {
+	defStore, err = store.NewDefaultStore(insertHdr, gj)
+	if err != nil {
+		return err
+	}
+
+	defQuery = query.NewDefaultQuery(selectHdr, gj)
+	defSubscriber = subscriber.NewDefaultSubscriber(topSub, varSub, defStore)
+	defService = service.NewService(defQuery)
+
+	return nil
+}
+
+func HTTPService(g *gin.RouterGroup) {
+	defService.HTTPService(g)
 }
 
 func Stop() {
-	subscriber.Stop()
-	store.Stop()
-	query.Stop()
+	defSubscriber.Close()
+	defQuery.Close()
+	defStore.Close()
 }
